pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
colorama = "0.4.6"
iniconfig = "2.0.0"
jmespath = "1.0.1"
packaging = "23.2"
pluggy = "1.3.0"
pytest = "7.4.3"
pyyaml = "^6.0.1"
pandas = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
